# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_kishert-lab_taxi-platform_internal_dto.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOi...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthConfirmCodeRequest:
    properties:
      code:
        example: "1234"
        maxLength: 6
        minLength: 4
        type: string
      name:
        example: Сергей
        maxLength: 255
        type: string
      phone:
        example: "+79124966126"
        type: string
    required:
    - code
    - phone
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOi...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOi...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeRequest:
    properties:
      phone:
        example: "+79124966126"
        type: string
    required:
    - phone
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeResponse:
    properties:
      message:
        example: Код подтверждения отправлен
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOi...
        type: string
      expires_in:
        example: 900
        type: integer
      passenger:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse'
      refresh_token:
        example: eyJhbGciOi...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse:
    properties:
      avatar_url:
        example: https://cdn.example.com/passengers/111/avatar.jpg
        type: string
      email:
        example: test@example.com
        type: string
      id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      name:
        example: Сергей
        type: string
      phone:
        example: "+79124966126"
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOi...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_maps.Configuration:
    properties:
      attribution:
        type: string
      bounds:
        items:
          type: number
        type: array
      data_version:
        type: string
      max_zoom:
        type: integer
      min_zoom:
        type: integer
      routing_available:
        type: boolean
      search_available:
        type: boolean
      style_url:
        type: string
      styles:
        items:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.Style'
        type: array
      updated_at:
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_maps.ReverseResponse:
    properties:
      address:
        type: string
      address_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
      found:
        type: boolean
      point:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
      source:
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_maps.RouteRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
        type: array
    required:
    - points
    type: object
  github_com_kishert-lab_taxi-platform_internal_maps.Style:
    properties:
      id:
        type: string
      url:
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_routing.Route:
    properties:
      calculated_at:
        type: string
      data_version:
        type: string
      distance_meters:
        type: integer
      duration_seconds:
        type: integer
      geometry:
        type: object
      original_points:
        items:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
        type: array
      snapped_points:
        items:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
        type: array
      source:
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.Error:
    properties:
      error:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorCode'
        example: ORDER_NOT_FOUND
      details:
        additionalProperties: {}
        type: object
      message:
        example: Order not found
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.ErrorCode:
    enum:
    - VALIDATION_ERROR
    - INVALID_PHONE
    - INVALID_CODE
    - CODE_EXPIRED
    - TOO_MANY_ATTEMPTS
    - PASSENGER_BLOCKED
    - TOKEN_EXPIRED
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - ORDER_NOT_FOUND
    - ORDER_INVALID_STATE
    - DRIVER_NOT_AVAILABLE
    - ORDER_ALREADY_ASSIGNED
    - DISPATCH_IN_PROGRESS
    - RATE_LIMITED
    - CONSENT_REQUIRED
    - NOT_IMPLEMENTED
    - INTERNAL_ERROR
    type: string
    x-enum-varnames:
    - CodeValidationError
    - CodeInvalidPhone
    - CodeInvalidCode
    - CodeCodeExpired
    - CodeTooManyAttempts
    - CodePassengerBlocked
    - CodeTokenExpired
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeOrderNotFound
    - CodeOrderInvalidState
    - CodeDriverNotAvailable
    - CodeOrderAlreadyAssigned
    - CodeDispatchInProgress
    - CodeRateLimited
    - CodeConsentRequired
    - CodeNotImplemented
    - CodeInternalError
  github_com_kishert-lab_taxi-platform_pkg_response.Meta:
    properties:
      request_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
    type: object
...
  internal_transport_http_handler.AcceptedSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_handler.acceptedResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.MapConfigurationResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.Configuration'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.MapReverseResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.ReverseResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.MapRouteResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_routing.Route'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerAuthRefreshSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRefreshResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerAuthRequestCodeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerAuthTokenSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthTokenResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerLogoutSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_handler.passengerLogoutResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.acceptedResponse:
    properties:
      accepted:
        example: true
        type: boolean
    type: object
...
  internal_transport_http_handler.passengerLogoutResponse:
    properties:
      logged_out:
        example: true
        type: boolean
    type: object
info:
  contact:
    email: support@example.com
    name: Taxi Platform API Support
  description: Production-grade taxi backend API for small cities and rural areas.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taxi Platform API
  version: 0.1.0
paths:
...
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the provided refresh token.
      parameters:
      - description: Logout request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.AcceptedSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Logout mobile user
      tags:
      - auth
...
  /geocoder/reverse:
    get:
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapReverseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Reverse geocode without changing the selected pickup point
      tags:
      - maps
...
  /map/routes:
    post:
      consumes:
      - application/json
      description: Points use latitude/longitude fields. GeoJSON geometry uses [longitude,latitude].
        Original points are preserved; snapped points are separate. No live traffic.
        Does not grant access to order or driver coordinates; obtain those through
        existing authorized order and WebSocket APIs.
      parameters:
      - description: Exactly two points; order model has no intermediate stops
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.RouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Calculate a road route between two supplied points
      tags:
      - maps
...
  /passenger/address/reverse:
    get:
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapReverseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Reverse geocode without changing the selected pickup point
      tags:
      - maps
...
  /passenger/auth/confirm-code:
    post:
      consumes:
      - application/json
      description: Verifies the one-time code and returns passenger access and refresh
        tokens.
      parameters:
      - description: Passenger code confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthConfirmCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Confirm passenger authentication code
      tags:
      - passenger-auth
  /passenger/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Passenger refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerLogoutSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Logout passenger
      tags:
      - passenger-auth
  /passenger/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Passenger refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthRefreshSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Refresh passenger tokens
      tags:
      - passenger-auth
  /passenger/auth/request-code:
    post:
      consumes:
      - application/json
      description: Sends a one-time confirmation code to the passenger phone number.
      parameters:
      - description: Passenger code request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthRequestCodeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Request passenger authentication code
      tags:
      - passenger-auth
...
  /passenger/map/routes:
    post:
      consumes:
      - application/json
      description: Points use latitude/longitude fields. GeoJSON geometry uses [longitude,latitude].
        Original points are preserved; snapped points are separate. No live traffic.
        Does not grant access to order or driver coordinates; obtain those through
        existing authorized order and WebSocket APIs.
      parameters:
      - description: Exactly two points; order model has no intermediate stops
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.RouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Calculate a road route between two supplied points
      tags:
      - maps
...
  /public/map/config:
    get:
      description: Availability indicates a published, validated dataset, not a live
        SLA. Bounds are west,south,east,north; west > east crosses the antimeridian.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapConfigurationResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Shared MapLibre configuration for all clients
      tags:
      - maps
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer ` prefix, for example: Bearer eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"